add email templates

+++ run.py
@@ -27,7 +27,8 @@ def run(cases, pattern, report):
     # 测试用例回溯
     try:
         for files in os.listdir(cases):
-            os.remove(cases + files)
+            if os.path.isfile(files):
+                os.remove(cases + files)
     except PermissionError:
         pass
